Add 'trigger' key to MultipleChoice's schema
Although MultipleChoice has a lot of common with Options step, there needs to be 'trigger' property for its step unlike Options step. This is because, Options step's next trigger simple: 'trigger' is inside the 'options' field. Same can not be done for MultipleChoice step as next step depends on combination of choices.