Fix OptionsStep's problem when loaded from the cache
OptionsStep when loaded from the cache had a problem, where the chosen option could not be matched up with either of the Options that was loaded from the cache. This matching was done using `filter()` where `===` operator is used. This caused problem as `===` operator only checks for Object reference equality. This has been fixed with the `deep-copy` package where Object's inner value is compared.