Reassign validator & parser to current step when loaded from cache
When a current step which is a user step is loaded from cache, it loses all reference to its validator and parser function. This occurs as `stringify()` does not stringify functions. Thus, the `validator` and `parser` function is reassigned to user step from the provided `steps`. In not done so, it would reject some of the perfectly valid user input.