Fix bug for reloading at UserStep and Add tests for same at Option
Bug regarding reloading at `UserStep` is that, when the chat is reloaded, and the steps are extracted from the `localStorage` cache, `parser` and `validator` is not assigned to the `currentStep` which is a `UserStep`.

Same bug resided when Chat is reloaded at `OptionStep`. But, that has already been solved by earlier commit. This commit just adds tests for it.